backend/requirements.txt

requests==2.31.0
python-dotenv==1.0.0
langchain==0.3.7
langchain-core==0.3.17
langchain-community==0.3.7
langchain-google-genai==2.0.5
langsmith==0.1.125
langchain-text-splitters==0.3.2
google-ai-generativelanguage==0.6.15
google-generativeai==0.8.5
pydantic-settings>=2.4.0,<3.0.0
